fix(edge-client): extract proxyFetchLegacy to lightweight cors-proxy subpath to avoid pulling protobufjs into workerd bundles
Add `@dxos/edge-client/cors-proxy` subpath export that contains only
`proxyFetchLegacy` with zero heavy transitive dependencies. Callers in
`@dxos/assistant-toolkit` and `@dxos/plugin-crm` now import from this
subpath, keeping the workerd bundle free of protobufjs (which comes in
via `@dxos/protocols/proto` when importing from the main edge-client
entry point).
