+++ packages/plugins/plugin-crm/src/operations/attach-image.ts
@@ -6,7 +6,7 @@ import * as Effect from 'effect/Effect';
 
 import { Operation } from '@dxos/compute';
 import { Database, Entity, Obj } from '@dxos/echo';
-import { proxyFetchLegacy } from '@dxos/edge-client';
+import { proxyFetchLegacy } from '@dxos/edge-client/cors-proxy';
 import { log } from '@dxos/log';
 import { Organization, Person } from '@dxos/types';
